Added warn() function with parameters matching die (and printf etc)
Changed coverage datatype to uint32_t throughout cortex_var
- All addition of coverages now check for overflow
- An error is printed (only once) if overflowing is seen

+++ scripts/analyse_variants/needleman_wunsch/libs/alignment_scoring/alignment.c
@@ -51,7 +51,7 @@ void alignment_print_matrices(score_t* match_score, score_t* gap_a_score,
     printf("%3i:", j);
     for(i = 0; i <= length_a; i++)
     {
-      printf(" %3i", ARR_LOOKUP(match_score, score_width, i, j));
+      printf(" %3i", (int)ARR_LOOKUP(match_score, score_width, i, j));
     }
     printf("\n");
   }
@@ -61,7 +61,7 @@ void alignment_print_matrices(score_t* match_score, score_t* gap_a_score,
     printf("%3i:", j);
     for(i = 0; i <= length_a; i++)
     {
-      printf(" %3i", ARR_LOOKUP(gap_a_score, score_width, i, j));
+      printf(" %3i", (int)ARR_LOOKUP(gap_a_score, score_width, i, j));
     }
     printf("\n");
   }
@@ -71,7 +71,7 @@ void alignment_print_matrices(score_t* match_score, score_t* gap_a_score,
     printf("%3i:", j);
     for(i = 0; i <= length_a; i++)
     {
-      printf(" %3i", ARR_LOOKUP(gap_b_score, score_width, i, j));
+      printf(" %3i", (int)ARR_LOOKUP(gap_b_score, score_width, i, j));
     }
     printf("\n");
   }
